Working out failure behavior.
The process that call ifastcgi:add_server/4 will be linked to by
the started server. This makes sure resources are freed.
The started server is also added as a temporary worker in the
supervisor which means it will not be restarted by the supervisor
on failure.
The idea is that we bring down the application that started the
server instead.